requirements.txt
playwright==1.55.0
python-dotenv==1.0.0